fix: change rating type back to Integer for database compatibility
- Rating stored as 1-10 (frontend displays as 5-star with half)
- Fixes 500 error caused by Float/Integer type mismatch in PostgreSQL

+++ backend/src/models/wine_item.py
@@ -75,7 +75,7 @@ class WineItem(Base):
 
     # 品飲筆記（喝完時填寫）
     tasting_notes = Column(String(1000), nullable=True)  # 舊欄位，保留相容
-    rating = Column(Float, nullable=True)  # 評分 (5星制，可半星)
+    rating = Column(Integer, nullable=True)  # 評分 (1-10，前端顯示為5星半星制)
     review = Column(String(1000), nullable=True)  # 評價
     flavor_tags = Column(String(500), nullable=True)  # 風味標籤 JSON array
     aroma = Column(String(500), nullable=True)  # 香氣
